Switch pane functions soon provided by shared
The internal structure of the shared state or the actions to change it
are not part of the public API of shared. So this commit switches to
new functions that actually will be part of the API.
