+++ src/reducers/appReducer.js
@@ -99,11 +99,6 @@ export const setFileLoadedAction = loaded => ({
     loaded,
 });
 
-export const setCurrentPane = currentPane => ({
-    type: SET_CURRENT_PANE,
-    currentPane,
-});
-
 export default (state = initialState, { type, ...action }) => {
     switch (type) {
         case DEVICE_OPENED: {
@@ -155,5 +150,3 @@ export const appState = ({ app }) => app.app;
 export const advancedMode = state => state.app.app.advancedMode;
 export const deviceOpen = state =>
     Object.keys(state.app.app.capabilities).length > 0;
-
-export const currentPane = state => state.appLayout.currentPane;
