+++ src/reducers/chartReducer.js
@@ -34,6 +34,7 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+import { currentPane } from '../from_pc-nrfconnect-shared';
 import { options } from '../globals';
 import { isRealTimePane } from '../utils/panes';
 
@@ -154,8 +155,9 @@ export const chartWindowAction = (
     yMax
 ) => (dispatch, getState) => {
     const { sampleFreq, maxSampleFreq } = getState().app.dataLogger;
-    const { currentPane } = getState().appLayout;
-    const sf = isRealTimePane(currentPane) ? maxSampleFreq : sampleFreq;
+    const sf = isRealTimePane(currentPane(getState()))
+        ? maxSampleFreq
+        : sampleFreq;
     const duration =
         windowDuration === null
             ? null
